Rename WorkspaceView to Workspace and reorganize module
